Minor cleanup in pending update tree comparison
Agent-Logs-Url: https://github.com/DanexCodr/Coderive/sessions/e09efce3-65e4-4b52-b570-1f833e7bbc72

+++ src/main/java/cod/range/NaturalArray.java
@@ -1238,7 +1238,7 @@ public class NaturalArray {
 
             if (winner != null && pendingUpdateOrderPrefixByStart != null) {
                 Long remainingMaxOrder = pendingUpdateOrderPrefixByStart.get(entry.getKey());
-                if (remainingMaxOrder != null && winner.order >= remainingMaxOrder.longValue()) {
+                if (remainingMaxOrder != null && winner.order >= remainingMaxOrder) {
                     break;
                 }
             }
